add helper struct for gamescope resolution
Contrary to what I thought, QMaps sort themselves automatically in
alphabetical order based off key.

This was resulting in the arguments coming in the opposite order they
would normally be input, ie "-h 1080 -w 1920", and meant that to keep
ordering consistent with expected order, we were iterating through the
QMap backwards.

This appears to have caused an issue with fullscreen gamescope, so this
commit changes it to be a temp struct that holds this information
instead.